[master] changed type data

+++ src/main/java/com/example/entity/Corresponsal.groovy
@@ -2,10 +2,11 @@ package com.example.entity
 
 import io.quarkus.hibernate.orm.panache.PanacheEntity
 
+import javax.persistence.Column
 import javax.persistence.Entity
 
 @Entity
 class Corresponsal  extends PanacheEntity {
-  Integer id
+  Long id
   String nombre
 }
